Add API class for encapsulating back-end
This class can be initialized with a username / password
combo, which it uses to get a token and stores that token.
This token will be used for all other methods / requests
it provides.

Dart does not support async factory constructors (see https://github.com/dart-lang/sdk/issues/23115)
so we use a static method instead. What is interesting is
that since all of this depends on a Network request, all
the steps in the pipeline return a Future. However, at
some point, we're going to need an actual instance before
we can proceed. Will be interesting to see how we get there.