Pass job instead of api to Analysis screen
Since the Analysis screen won't be making any API calls,
it doesn't make sense to give it the entire API. It only
needs the Future Jobs to show what it needs.

Currently it shows a Circular Progress Indicator on the
first page when loading, a checkmark when done, or an
error icon in case of problems.